backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --frozen --no-hashes
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
click==8.3.0
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
fastapi==0.119.1
    # via backend
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via anyio
psycopg2-binary==2.9.11
    # via backend
pydantic==2.12.3
    # via fastapi
pydantic-core==2.41.4
    # via pydantic
python-dotenv==1.1.1
    # via backend
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via backend
starlette==0.48.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.38.0
    # via backend
